FEAT: sigun completed and working with composable
