+++ src/composables/views/useAuthForm.composable.ts
@@ -3,13 +3,9 @@ import { ref, computed } from 'vue'
 export function useAuthForm() {
   const email = ref<string>('')
   const password = ref<string>('')
-  const name = ref<string>('')
-  const lastname = ref<string>('')
   const alertMessage = ref<string>('')
   const emailErrors = ref<string[]>([])
   const passwordErrors = ref<string[]>([])
-  const nameErrors = ref<string[]>([])
-  const lastnameErrors = ref<string[]>([])
 
   const displayAlert = computed(() => alertMessage.value.length > 0)
   const isDisabled = computed(() => {
@@ -17,9 +13,7 @@ export function useAuthForm() {
       emailErrors.value.length > 0 ||
       passwordErrors.value.length > 0 ||
       email.value.length === 0 ||
-      password.value.length === 0 ||
-      name.value.length === 0 ||
-      lastname.value.length === 0
+      password.value.length === 0
     )
   })
 
@@ -41,24 +35,6 @@ export function useAuthForm() {
     }
   }
 
-  function handleName({ value, isValid }: { value: string; isValid: boolean }): void {
-    name.value = value
-    if (!isValid) {
-      nameErrors.value = ['Campo obligatorio']
-    } else {
-      nameErrors.value = []
-    }
-  }
-
-  function handleLastname({ value, isValid }: { value: string; isValid: boolean }): void {
-    lastname.value = value
-    if (!isValid) {
-      lastnameErrors.value = ['Campo obligatorio']
-    } else {
-      lastnameErrors.value = []
-    }
-  }
-
   function closeAlert(): void {
     alertMessage.value = ''
   }
@@ -67,7 +43,6 @@ export function useAuthForm() {
     if (emailErrors.value.length === 0 && passwordErrors.value.length === 0) {
       console.log('Email:', email.value)
       console.log('Password:', password.value)
-      console.log('Name:', name.value)
     }
     alertMessage.value = 'Corrige los errores antes de continuar'
   }
@@ -75,8 +50,6 @@ export function useAuthForm() {
   return {
     email,
     password,
-    name,
-    lastname,
     alertMessage,
     emailErrors,
     passwordErrors,
@@ -84,8 +57,6 @@ export function useAuthForm() {
     isDisabled,
     handleEmailValidation,
     handlePasswordValidation,
-    handleName,
-    handleLastname,
     closeAlert,
     submitForm
   }
